# contracts/api.openapi.yaml
openapi: 3.0.3
info:
  title: PetriDishOfMadness Control API
  version: 1.2.0
  description: >
    REST control + config surface for the chaos lab. The live data stream is WS /ws
    (see events.schema.json). All endpoints are unauthenticated (local-first, single-user).
    v1.1.0 (W6) adds read-only endpoints over the append-only event log (event-log.md §7)
    that power the /inspector annex: events, turn traces, rule history, relationships,
    snapshots, replay, and analytics (9-AWI + model-vs-model). These serve deep/full-history
    replay in live mode; the /inspector panels also compute from the client-side rolling
    history so they render in mock mode with no backend.
servers:
  - url: http://localhost:8000
paths:
  /api/health:
    get:
      summary: Liveness
      responses: { "200": { description: ok, content: { application/json: { schema: { type: object, properties: { status: { type: string }, tick: { type: integer }, running: { type: boolean } } } } } } }
  /api/state:
    get:
      summary: Full world snapshot (same shape as WS world_state)
      responses: { "200": { description: snapshot, content: { application/json: { schema: { type: object } } } } }
  /api/config:
    get:
      summary: Effective world config
      responses: { "200": { description: config, content: { application/json: { schema: { type: object } } } } }
  /api/profiles:
    get:
      summary: Model profiles (the legend) with availability
      responses: { "200": { description: profiles, content: { application/json: { schema: { type: array, items: { type: object, properties: { name: {type: string}, adapter: {type: string}, model_id: {type: string}, color: {type: string}, available: {type: boolean} } } } } } } }
  /api/control/start:
    post: { summary: Start/resume the loop, responses: { "200": { description: ok } } }
  /api/control/pause:
    post: { summary: Pause the loop, responses: { "200": { description: ok } } }
  /api/control/step:
    post: { summary: Advance exactly one turn, responses: { "200": { description: ok } } }
  /api/control/speed:
    post:
      summary: Set tick interval (seconds between turns)
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [tick_interval_seconds], properties: { tick_interval_seconds: { type: number, minimum: 0.1, maximum: 60 } } } } } }
      responses: { "200": { description: ok } }
  /api/control/reset:
    post:
      summary: Reset the world from config (optionally override seed)
      responses: { "200": { description: ok } }
  /api/agents/{id}/model:
    post:
      summary: Reassign an agent's model profile (takes effect next turn) — the marquee feature
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [profile], properties: { profile: { type: string } } } } } }
      responses: { "200": { description: ok }, "404": { description: unknown agent }, "400": { description: unknown profile } }
  /api/agents:
    post:
      summary: Spawn an agent (W7 EM-063 — god-mode immediate, or governance-gated when mode=governance)
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name, profile], properties: { name: {type: string}, personality: {type: string}, profile: {type: string}, location: {type: string}, mode: { type: string, enum: [god, governance], default: god, description: "god=immediate; governance=enqueue admit_agent proposal" } } } } } }
      responses: { "201": { description: "created (god) — emits agent_spawned{method:god}" }, "202": { description: "accepted (governance) — admit_agent proposal enqueued; agent admitted iff vote passes" } }
  /api/animals:
    get:
      summary: List animals with state (W8 EM-064) — also present in world_state.animals; empty array when no run/animals (never 500)
      responses: { "200": { description: animals, content: { application/json: { schema: { type: array, items: { type: object } } } } } }
    post:
      summary: Spawn an animal immediately (god-mode), emitting animal_spawned. Animals have no credits and are not in the agent round-robin.
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [species, name], properties: { species: { type: string, enum: [cat, dog] }, name: { type: string }, location: { type: string, default: plaza }, personality: { type: string, default: "" } } } } } }
      responses: { "201": { description: "created — {status, animal_id}" }, "400": { description: unknown species }, "503": { description: not initialized } }
  /api/buildings:
    get:
      summary: List buildings/structures with state (W7 EM-061) — also available in world_state.buildings
      responses: { "200": { description: buildings, content: { application/json: { schema: { type: array, items: { type: object } } } } } }
  /api/agents/{id}:
    delete:
      summary: Kill/remove an agent
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { description: ok } }
  /api/events/inject:
    post:
      summary: Inject a random world event (chaos knob)
      requestBody: { content: { application/json: { schema: { type: object, properties: { kind: { type: string, enum: [windfall, famine, blackout, festival], description: "optional; random if omitted" } } } } } }
      responses: { "200": { description: ok } }
  # --- W6 read-only event-log endpoints (event-log.md §7). Run-scoped to the active run. ---
  /api/events:
    get:
      summary: Query the append-only event log
      parameters:
        - { name: from_tick, in: query, schema: { type: integer } }
        - { name: to_tick, in: query, schema: { type: integer } }
        - { name: kinds, in: query, schema: { type: string }, description: "comma-separated kinds" }
        - { name: actor_id, in: query, schema: { type: string } }
        - { name: turn_id, in: query, schema: { type: string } }
        - { name: after_seq, in: query, schema: { type: integer }, description: "keyset pagination (seq > after_seq)" }
        - { name: limit, in: query, schema: { type: integer } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc], default: asc } }
      responses: { "200": { description: events, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/EventRow' } } } } } }
  /api/turns/{turn_id}:
    get:
      summary: The full ordered decision-trace chain for one turn (EM-056)
      parameters: [ { name: turn_id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { description: chain, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/EventRow' } } } } } }
  /api/rules/history:
    get:
      summary: Governance/laws history — proposal lifecycle + downstream consequences (EM-057)
      responses: { "200": { description: rules, content: { application/json: { schema: { type: array, items: { type: object } } } } } }
  /api/relationships:
    get:
      summary: Relationship/conflict/gift events for the social graph, time-scrubbable (EM-058)
      parameters:
        - { name: agent_id, in: query, schema: { type: string } }
        - { name: from_tick, in: query, schema: { type: integer } }
        - { name: to_tick, in: query, schema: { type: integer } }
      responses: { "200": { description: timeline, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/EventRow' } } } } } }
  /api/snapshots:
    get:
      summary: List snapshot ticks (ascending)
      responses: { "200": { description: ticks, content: { application/json: { schema: { type: array, items: { type: object, properties: { tick: { type: integer } } } } } } } }
  /api/replay:
    get:
      summary: Replay materials for tick T — nearest prior snapshot + the fold-forward delta (EM-055)
      description: >
        v1.2.0 (W9, audit §B7): `events` contains ONLY the delta — rows with
        base.tick < event.tick <= T (strict on the left; the snapshot at base.tick
        already includes all tick-base.tick events, per event-log.md v1.1.0 §boundary).
        Clients fold `events` onto `base.state` without further filtering.
      parameters: [ { name: tick, in: query, required: true, schema: { type: integer } } ]
      responses: { "200": { description: replay, content: { application/json: { schema: { type: object, properties: { base: { type: object, properties: { tick: { type: integer }, state: { type: object } } }, events: { type: array, items: { $ref: '#/components/schemas/EventRow' } } } } } } } }
  /api/analytics:
    get:
      summary: 9-AWI + model-vs-model analytics (event-log.md §7 get_analytics shape) (EM-059/067)
      parameters:
        - { name: from_tick, in: query, schema: { type: integer } }
        - { name: to_tick, in: query, schema: { type: integer } }
      responses: { "200": { description: analytics, content: { application/json: { schema: { type: object } } } } }
  /ws:
    get:
      summary: WebSocket event stream (see events.schema.json). Documented here for discoverability.
      responses: { "101": { description: switching protocols } }
components:
  schemas:
    EventRow:
      type: object
      description: One append-only event-log row (event-log.md §2/§7). payload is kind-specific.
      properties:
        seq: { type: integer }
        run_id: { type: integer }
        tick: { type: integer }
        sim_time: { type: number, nullable: true }
        kind: { type: string }
        actor_id: { type: string, nullable: true }
        actor_type: { type: string }
        target_id: { type: string, nullable: true }
        profile: { type: string, nullable: true }
        turn_id: { type: string, nullable: true }
        text: { type: string, nullable: true }
        payload: { type: object }
        ts: { type: string }